+++ src/routing/Chapter.tsx
@@ -1,5 +1,11 @@
-import React, { ReactElement } from 'react';
-import { createRoutesFromChildren, Link, Outlet } from 'react-router-dom';
+import React, { ReactElement, useContext } from 'react';
+import {
+  createRoutesFromChildren,
+  Link,
+  Outlet,
+  useLocation,
+} from 'react-router-dom';
+import { RouterContext } from './RouterContext';
 import { ChapterProps } from './types';
 
 /**
@@ -8,6 +14,9 @@ import { ChapterProps } from './types';
  * @beta
  */
 export default function Chapter(props: ChapterProps): JSX.Element {
+  const listOfRoutes = useContext(RouterContext).listOfRoutes;
+  const currentLocation = useLocation();
+
   return (
     <div className="chapter">
       {props.title && <h2>{props.title}</h2>}
