+++ src/routing/Page.tsx
@@ -8,7 +8,7 @@ import {
   RouteObject,
 } from 'react-router-dom';
 import { PageProps, IFormData } from './types';
-import { RouterContext } from './Router';
+import { RouterContext } from './RouterContext';
 
 /**
  * Renders the page contents
@@ -18,11 +18,8 @@ import { RouterContext } from './Router';
 export default function Page(props: PageProps): JSX.Element {
   const { values, submitForm } = useFormikContext();
   const formValues = values as IFormData;
-  const routePages = useContext(RouterContext).listOfRoutes;
+  const listOfRoutes = useContext(RouterContext).listOfRoutes;
   const currentLocation = useLocation();
-
-  // get the current route in order to determine the next route.
-  const splitLocation = currentLocation.pathname.split('/');
 
   const navigate = useNavigate();
